+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+# Next release
+
 # Release v0.4.0
 
 ## Highlights